Refactoring bolton package to bolt_on only in code usages.
