+++ privacy/bolt_on/__init__.py
@@ -23,7 +23,7 @@ if LooseVersion(tf.__version__) < LooseVersion("2.0.0"):
 if hasattr(sys, "skip_tf_privacy_import"):  # Useful for standalone scripts.
   pass
 else:
-  from privacy.bolton.models import BoltOnModel  # pylint: disable=g-import-not-at-top
-  from privacy.bolton.optimizers import BoltOn  # pylint: disable=g-import-not-at-top
-  from privacy.bolton.losses import StrongConvexHuber  # pylint: disable=g-import-not-at-top
-  from privacy.bolton.losses import StrongConvexBinaryCrossentropy  # pylint: disable=g-import-not-at-top
+  from privacy.bolt_on.models import BoltOnModel  # pylint: disable=g-import-not-at-top
+  from privacy.bolt_on.optimizers import BoltOn  # pylint: disable=g-import-not-at-top
+  from privacy.bolt_on.losses import StrongConvexHuber  # pylint: disable=g-import-not-at-top
+  from privacy.bolt_on.losses import StrongConvexBinaryCrossentropy  # pylint: disable=g-import-not-at-top
